requirements.txt
streamlit == 1.32.0
openai == 1.12.0
python-dotenv ==1.0.1
sounddevice == 0.4.6
numpy == 1.26.4
